Changed the code structure.
